Fix a few review comments
